Improve resizing behavior
This commit tweaks the various stylesheets so that Grimoire will behave
better in terms of not reducing font sizes and in terms of rearranging
content on smaller screens. This should make mobile Grimoire not totally
useless. It also means that when you shrink a Grimoire window the font
won't drop on you.