Versions freeze
"make versions" was intended for 0.2.X when the monolithic nature of the data
files required that they be blown away relatively regularly. As this is no
longer the case in 0.3.0 thanks to the introduction of modular examples among
other changes, the versions target is not only unneeded, it's dangerous.
Best to just kill it off.