Make kw/symbol/comment syntax highlighting more visible
Keywords/symbols used to be a light yellow that was hard to read on
the light background, changed to use something almost equal to the
yellow from Solarized.

For comments, the current grey would be fine on a white bg, but on
the light grey code example bg, it needs to be slightly darker for
contrast.  It could go even darker, to #808080 or so.