Decomplect URL upgrading code from rendering code
This patch does a shitload of work, in that it in one stroke cleans up
all the old URL redirecting code by unwinding it from the `store` route
into inline pattern matches in the main `app` route.

This patch also versions the store URL so that future route
upgrades (god forbid) will be easier to implement.