Use short things in the database not custom strs
This fixes clojure-grimoire/grimoire#195

Sorta kinda fixed. For aggregating "heat" information (frequency of
access) we care about frequency on logical structures like namespaces
and defs not specific Thing instances. As a result, the right structure
to use for this is the lossy "short thing" representation since Things
regardless of exact artifact and version all "hash" to the same string
using this representation. The short thing structure provides a way to
round trip from one logical Thing to "the" logical Thing that it
_probably_ came from in the first place which I consider good enough.