Implement a render cache for the worklist page
Fixes a bug exposed by @danielcompton where the Grimoire host can be
brought literally to its knees by repeated requests for the
worklist. This isn't a perfect fix, the caches should be refreshed on
some clock tick, but since right now a full restart is required for a
data file update this is good enough.